Fix issue related to usage of manual loading of annotations
Gedmo uses the AnnotationRegistry::registerFile which toggles the bool AnnotationRegistry$registerFileUsed.
This causes AnnotationRegistry::loadAnnotationClass() to go into a mode where it expects all annotations to be loaded in a different manner, and not autoloaded.
This behaviour from Gedmo was changed here doctrine-extensions/DoctrineExtensions#2558

Adding this custom loader makes the annotation registry "autoloadable".